export HeroContentProps
